@@ -31,14 +30,14 @@
         <dep.docker-java.version>3.2.12</dep.docker-java.version>
         <dep.coral.version>2.0.55</dep.coral.version>
         <dep.confluent.version>5.5.2</dep.confluent.version>
-        <project.build.targetJdk>17</project.build.targetJdk>
-        <air.java.version>1.8.0-202</air.java.version>
-        <dep.trino.version>429</dep.trino.version>
-        <dep.trino.airlift.version>236</dep.trino.airlift.version>
+        <project.build.targetJdk>22</project.build.targetJdk>
+        <air.java.version>22.0.0</air.java.version>
+        <dep.trino.version>462</dep.trino.version>
+        <dep.trino.airlift.version>248</dep.trino.airlift.version>
         <dep.packaging.version>${dep.trino.airlift.version}</dep.packaging.version>
-        <dep.arrow.version>13.0.0</dep.arrow.version>
-        <maven.compiler.source>17</maven.compiler.source>
-        <maven.compiler.target>17</maven.compiler.target>
+        <dep.arrow.version>17.0.0</dep.arrow.version>
+        <maven.compiler.source>22</maven.compiler.source>
+        <maven.compiler.target>22</maven.compiler.target>
 
         <dep.docker.images.version>53</dep.docker.images.version>
 
@@ -446,8 +455,8 @@
                 <groupId>org.apache.maven.plugins</groupId>
                 <artifactId>maven-compiler-plugin</artifactId>
                 <configuration combine.self="override">
-                    <source>17</source>
-                    <target>17</target>
+                    <source>22</source>
+                    <target>22</target>
                     <fork>false</fork>
                 </configuration>
             </plugin>
