@@ -5,8 +5,6 @@
 package spark.sql.catalog.ndb;
 
 import com.google.common.collect.ImmutableList;
-import com.google.common.collect.ImmutableSet;
-import org.apache.arrow.memory.ArrowBuf;
 import org.apache.arrow.vector.BigIntVector;
 import org.apache.arrow.vector.BitVector;
 import org.apache.arrow.vector.DateDayVector;
@@ -26,7 +24,6 @@ import org.apache.arrow.vector.VectorSchemaRoot;
 import org.apache.arrow.vector.complex.ListVector;
 import org.apache.arrow.vector.complex.MapVector;
 import org.apache.arrow.vector.complex.StructVector;
-import org.apache.arrow.vector.ipc.message.ArrowFieldNode;
 import org.apache.arrow.vector.types.TimeUnit;
 import org.apache.arrow.vector.types.pojo.ArrowType;
 import org.apache.arrow.vector.types.pojo.Field;
@@ -48,36 +45,39 @@ import org.apache.spark.sql.execution.arrow.ShortWriter;
 import org.apache.spark.sql.execution.arrow.StringWriter;
 import org.apache.spark.sql.execution.arrow.StructWriter;
 import org.apache.spark.sql.execution.arrow.TimestampWriter;
+import org.apache.spark.sql.types.ArrayType;
 import org.apache.spark.sql.types.CharType;
 import org.apache.spark.sql.types.DataType;
 import org.apache.spark.sql.types.DataTypes;
+import org.apache.spark.sql.types.MapType;
 import org.apache.spark.sql.types.Metadata;
 import org.apache.spark.sql.types.MetadataBuilder;
 import org.apache.spark.sql.types.StructField;
 import org.apache.spark.sql.types.StructType;
 import org.apache.spark.sql.types.VarcharType;
 import org.apache.spark.sql.util.ArrowUtils;
+import org.jetbrains.annotations.NotNull;
+import scala.Function1;
 
 import java.util.Arrays;
 import java.util.List;
-import java.util.Set;
 import java.util.stream.Collectors;
 import java.util.stream.IntStream;
 
 import static java.lang.String.format;
-import static org.apache.arrow.vector.complex.BaseRepeatedValueVector.OFFSET_WIDTH;
 import static org.apache.spark.sql.types.DataTypes.TimestampType;
 import static org.apache.spark.sql.types.DataTypes.createStructField;
+import static spark.sql.catalog.ndb.TimestampPrecision.MICROSECONDS;
+import static spark.sql.catalog.ndb.UnsupportedTypes.UNSUPPORTED_TYPE_PREDICATE;
 
 public final class TypeUtil
 {
     public static final String NDB_CATALOG_DOES_NOT_SUPPORT_NON_NULL_COLUMNS = "NDB catalog does not support Non null columns";
     public static final String NDB_CATALOG_DOES_NOT_SUPPORT_TYPES = "NDB catalog does not support the following types:";
 
-    // possible types from ArrowUtils.toArrowType that are unsupported by vast
-    private static final Set<ArrowType.ArrowTypeID> unsupportedNDBTypes = ImmutableSet.of(ArrowType.ArrowTypeID.Interval,
-            ArrowType.ArrowTypeID.Duration,
-            ArrowType.ArrowTypeID.Null);
+    // Arrow-specific column names for representing nested data types
+    static final String ARRAY_ITEM_COLUMN_NAME = "item";
+
     public static final String TIMESTAMP_PRECISION = "TIMESTAMP_PRECISION";
 
     private TypeUtil() {}
@@ -90,7 +90,7 @@ public final class TypeUtil
         }
 
 
-        List<Field> unsupportedFields = fieldList.stream().filter(field -> unsupportedNDBTypes.contains(field.getType().getTypeID())).collect(Collectors.toList());
+        List<Field> unsupportedFields = fieldList.stream().filter(UNSUPPORTED_TYPE_PREDICATE).collect(Collectors.toList());
         if (!unsupportedFields.isEmpty()) {
             throw new UnsupportedOperationException(format("%s %s", NDB_CATALOG_DOES_NOT_SUPPORT_TYPES, unsupportedFields));
         }
@@ -116,29 +116,39 @@ public final class TypeUtil
 
     public static StructField arrowFieldToSparkField(Field arrowField)
     {
+        MetadataBuilder metadataBuilder = new MetadataBuilder();
         DataType dataType;
         try {
-            if (arrowField.getType().getTypeID() == ArrowType.ArrowTypeID.Timestamp) {
-                MetadataBuilder metadataBuilder = new MetadataBuilder();
+            if (arrowField.getType().getTypeID() == ArrowType.ArrowTypeID.List) {
+                Field childField = arrowField.getChildren().get(0);
+                StructField childStructField = arrowFieldToSparkField(childField);
+                dataType = new ArrayType(childStructField.dataType(), childStructField.nullable());
+            }
+            else if (arrowField.getType().getTypeID() == ArrowType.ArrowTypeID.Map) {
+                Field structSubField = arrowField.getChildren().get(0);
+                List<Field> mapFields = structSubField.getChildren();
+                Field keyField = mapFields.get(0);
+                Field valField = mapFields.get(1);
+                StructField keyStructField = arrowFieldToSparkField(keyField);
+                StructField valStructField = arrowFieldToSparkField(valField);
+                dataType = new MapType(keyStructField.dataType(), valStructField.dataType(), valStructField.nullable());
+            }
+            else if (arrowField.getType().getTypeID() == ArrowType.ArrowTypeID.Struct) {
+                StructField[] children = arrowField.getChildren().stream()
+                        .map(TypeUtil::arrowFieldToSparkField).toArray(StructField[]::new);
+                dataType = new StructType(children);
+            }
+            else if (arrowField.getType().getTypeID() == ArrowType.ArrowTypeID.Timestamp) {
                 ArrowType.Timestamp type = (ArrowType.Timestamp) arrowField.getType();
                 TimeUnit unit = type.getUnit();
-                switch (unit) {
-                    case SECOND:
-                        metadataBuilder.putLong(TIMESTAMP_PRECISION, TimestampPrecision.SECONDS.getPrecisionID());
-                        break;
-                    case MILLISECOND:
-                        metadataBuilder.putLong(TIMESTAMP_PRECISION, TimestampPrecision.MILLISECONDS.getPrecisionID());
-                        break;
-                    case MICROSECOND:
-                        metadataBuilder.putLong(TIMESTAMP_PRECISION, TimestampPrecision.MICROSECONDS.getPrecisionID());
-                        break;
-                    case NANOSECOND:
-                        metadataBuilder.putLong(TIMESTAMP_PRECISION, TimestampPrecision.NANOSECONDS.getPrecisionID());
-                        break;
-                }
-                return createStructField(arrowField.getName(), TimestampType, arrowField.isNullable(), metadataBuilder.build());
+                int timeUnitPrecisionID = getTimeUnitPrecisionID(unit);
+                metadataBuilder.putLong(TIMESTAMP_PRECISION, timeUnitPrecisionID);
+                dataType = TimestampType;
             }
-            dataType = ArrowUtils.fromArrowField(arrowField);
+            else {
+                dataType = ArrowUtils.fromArrowField(arrowField);
+            }
+
         }
         catch (Exception exception) {
             if (arrowField.getType().getTypeID() == ArrowType.ArrowTypeID.FixedSizeBinary) {
@@ -147,7 +157,23 @@ public final class TypeUtil
                 throw exception;
             }
         }
-        return createStructField(arrowField.getName(), dataType, arrowField.isNullable());
+        return createStructField(arrowField.getName(), dataType, arrowField.isNullable(), metadataBuilder.build());
+    }
+
+    private static int getTimeUnitPrecisionID(TimeUnit unit)
+    {
+        switch (unit) {
+            case SECOND:
+                return TimestampPrecision.SECONDS.getPrecisionID();
+            case MILLISECOND:
+                return TimestampPrecision.MILLISECONDS.getPrecisionID();
+            case MICROSECOND:
+                return MICROSECONDS.getPrecisionID();
+            case NANOSECOND:
+                return TimestampPrecision.NANOSECONDS.getPrecisionID();
+            default:
+                throw new RuntimeException(format("Unexpected TimeUnit: %s", unit));
+        }
     }
 
     public static Field sparkFieldToArrowField(StructField structField)
@@ -157,48 +183,91 @@ public final class TypeUtil
 
     public static Field sparkFieldToArrowField(String name, DataType sparkType, boolean nullable, Metadata metadata)
     {
-        try {
-            if (sparkType.equals(DataTypes.TimestampType)) {
-                if (metadata.contains(TIMESTAMP_PRECISION)) {
-                    long aLong = metadata.getLong(TIMESTAMP_PRECISION);
-                    TimestampPrecision timestampPrecision = TimestampPrecision.fromID((int) aLong);
-                    if (timestampPrecision == null) {
-                        throw new RuntimeException(format("Unexpected prevision for type: %s, with metadata: %s",
-                                sparkType, metadata));
+        if (sparkType instanceof MapType) {
+            MapType mapType = (MapType) sparkType;
+            DataType keyType = mapType.keyType();
+            DataType valueType = mapType.valueType();
+            StructType structChildType = new StructType()
+                    .add(MapVector.KEY_NAME, keyType, false)
+                    .add(MapVector.VALUE_NAME, valueType, mapType.valueContainsNull());
+            Field structChildField = sparkFieldToArrowField(MapVector.DATA_VECTOR_NAME, structChildType, false, metadata);
+            ArrowType arrowType = new ArrowType.Map(false);
+            return new Field(name,
+                    nullable ? FieldType.nullable(arrowType) : FieldType.notNullable(arrowType),
+                    ImmutableList.of(structChildField));
+        }
+        else if (sparkType instanceof ArrayType) {
+            ArrayType arrayType = (ArrayType) sparkType;
+            DataType childType = arrayType.elementType();
+            Field arrayArrowField = sparkFieldToArrowField(ARRAY_ITEM_COLUMN_NAME, childType, true, metadata);
+            ArrowType arrowType = ArrowType.List.INSTANCE;
+            return new Field(name,
+                    nullable ? FieldType.nullable(arrowType) : FieldType.notNullable(arrowType),
+                    ImmutableList.of(arrayArrowField));
+        }
+        else if (sparkType instanceof StructType) {
+            StructType struct = (StructType) sparkType;
+            List<Field> children = IntStream.range(0, struct.size())
+                    .mapToObj(i -> {
+                        StructField structField = struct.apply(i);
+                        DataType type = structField.dataType();
+                        return sparkFieldToArrowField(structField.name(), type, structField.nullable(), metadata);
+                    })
+                    .collect(Collectors.toList());
+            ArrowType arrowType = ArrowType.Struct.INSTANCE;
+            return new Field(name,
+                    nullable ? FieldType.nullable(arrowType) : FieldType.notNullable(arrowType), children);
+        }
+        else {
+            try {
+                if (sparkType.equals(DataTypes.TimestampType)) {
+                    TimestampPrecision timestampPrecision = MICROSECONDS;
+                    if (metadata.contains(TIMESTAMP_PRECISION)) {
+                        long aLong = metadata.getLong(TIMESTAMP_PRECISION);
+                        timestampPrecision = TimestampPrecision.fromID((int) aLong);
+                        if (timestampPrecision == null) {
+                            throw new RuntimeException(format("Unexpected prevision for type: %s, with metadata: %s",
+                                    sparkType, metadata));
+                        }
                     }
-                    ArrowType arrowType;
-                    switch (timestampPrecision) {
-                        case SECONDS:
-                            arrowType = new ArrowType.Timestamp(TimeUnit.SECOND, "UTC");
-                            break;
-                        case MILLISECONDS:
-                            arrowType = new ArrowType.Timestamp(TimeUnit.MILLISECOND, "UTC");
-                            break;
-                        case MICROSECONDS:
-                            arrowType = new ArrowType.Timestamp(TimeUnit.MICROSECOND, "UTC");
-                            break;
-                        case NANOSECONDS:
-                            arrowType = new ArrowType.Timestamp(TimeUnit.NANOSECOND, "UTC");
-                            break;
-                        default:
-                            throw new RuntimeException(format("Unexpected timestamp type precision: %s for type: %s",
-                                    timestampPrecision, sparkType));
-                    }
-                    FieldType fieldType = new FieldType(nullable, arrowType, null);
+                    FieldType fieldType = getTimestampFieldType(sparkType, nullable, timestampPrecision);
                     return new Field(name, fieldType, null);
                 }
+                return ArrowUtils.toArrowField(name, sparkType, nullable, null); // TODO - check large var types support
             }
-            return ArrowUtils.toArrowField(name, sparkType, nullable, "UTC");
+            catch (UnsupportedOperationException exception) {
+                if (sparkType.sameType(new VarcharType(sparkType.defaultSize()))) {
+                    return new Field(name, new FieldType(nullable, ArrowType.Utf8.INSTANCE, null), null);
+                }
+                if (sparkType.sameType(new CharType(sparkType.defaultSize()))) {
+                    return new Field(name, new FieldType(nullable, new ArrowType.FixedSizeBinary(((CharType) sparkType).length()), null), null);
+                }
+                throw exception;
+            }
         }
-        catch (UnsupportedOperationException exception) {
-            if (sparkType.sameType(new VarcharType(sparkType.defaultSize()))) {
-                return new Field(name, new FieldType(true, ArrowType.Utf8.INSTANCE, null), null);
-            }
-            if (sparkType.sameType(new CharType(sparkType.defaultSize()))) {
-                return new Field(name, new FieldType(true, new ArrowType.FixedSizeBinary(((CharType) sparkType).length()), null), null);
-            }
-            throw exception;
+    }
+
+    private static @NotNull FieldType getTimestampFieldType(DataType sparkType, boolean nullable, TimestampPrecision timestampPrecision)
+    {
+        ArrowType arrowType;
+        switch (timestampPrecision) {
+            case SECONDS:
+                arrowType = new ArrowType.Timestamp(TimeUnit.SECOND, "UTC");
+                break;
+            case MILLISECONDS:
+                arrowType = new ArrowType.Timestamp(TimeUnit.MILLISECOND, "UTC");
+                break;
+            case MICROSECONDS:
+                arrowType = new ArrowType.Timestamp(TimeUnit.MICROSECOND, "UTC");
+                break;
+            case NANOSECONDS:
+                arrowType = new ArrowType.Timestamp(TimeUnit.NANOSECOND, "UTC");
+                break;
+            default:
+                throw new RuntimeException(format("Unexpected timestamp type precision: %s for type: %s",
+                        timestampPrecision, sparkType));
         }
+        return new FieldType(nullable, arrowType, null);
     }
 
     public static ArrowWriter getArrowSchemaWriter(VectorSchemaRoot root)
@@ -278,7 +347,7 @@ public final class TypeUtil
                     case MICROSECOND:
                         return new TimestampWriter((TimeStampMicroTZVector) vector);
                 }
-                throw new RuntimeException("Unexpected arrow timesatmp type: %s");
+                throw new RuntimeException("Unexpected arrow timestamp type: %s");
             }
             case Map: {
                 MapVector mapVector = (MapVector) vector;
@@ -304,32 +373,5 @@ public final class TypeUtil
         throw new IllegalArgumentException("unsupported Arrow type: " + arrowType);
     }
 
-    public static void convertFixedSizeBinaryIntoVarchar(FixedSizeBinaryVector src, VarCharVector dst)
-    {
-        final int valueCount = src.getValueCount();
-        final int byteWidth = src.getByteWidth();
-        ArrowBuf data = src.getDataBuffer();
-        ArrowBuf validity = src.getValidityBuffer();
-        try (ArrowBuf offsets = dst.getAllocator().buffer((long) (valueCount + 1) * OFFSET_WIDTH)) {
-            int offset = 0;
-            int index = 0;
-            for (int i = 0; i <= valueCount; ++i) {
-                offsets.setInt(index, offset);
-                index += OFFSET_WIDTH;
-                offset += byteWidth;
-            }
-            ArrowFieldNode node = new ArrowFieldNode(valueCount, src.getNullCount());
-            // See https://arrow.apache.org/docs/format/Columnar.html#buffer-listing-for-each-layout for buffer order definition
-            dst.loadFieldBuffers(node, ImmutableList.of(validity, offsets, data));
-            dst.setValueCount(valueCount);
-        }
-    }
-
-    public static void convertNonMicroTSVectorToMicroTSVector(TimeStampVector src, TimeStampMicroTZVector dst)
-    {
-        int valueCount = src.getValueCount();
-        List<ArrowBuf> fieldBuffers = src.getFieldBuffers();
-        ArrowFieldNode node = new ArrowFieldNode(valueCount, src.getNullCount());
-        dst.loadFieldBuffers(node, fieldBuffers);
-    }
+    public static final Function1<DataType, Object> schemaHasCharNType = type -> type instanceof CharType;
 }
