@@ -16,19 +16,21 @@ public final class MockMapSchema
 {
     private final String name;
     private final Map<String, MockTable> tables;
+    private final Map<String, MockView> views;
     private final Map<String, String> properties;
 
-    private MockMapSchema(String name, Map<String, MockTable> tables, Map<String, String> properties)
+    private MockMapSchema(String name, Map<String, MockTable> tables, Map<String, MockView> views, Map<String, String> properties)
     {
         this.name = name;
         this.tables = tables;
+        this.views = views;
         this.properties = properties;
     }
 
     public static MockMapSchema empty(String name)
     {
         checkArgument(!Strings.isNullOrEmpty(name));
-        return new MockMapSchema(name, new HashMap<>(), new HashMap<>());
+        return new MockMapSchema(name, new HashMap<>(), new HashMap<>(), new HashMap<>());
     }
 
     public String getName()
@@ -39,6 +41,11 @@ public final class MockMapSchema
     public Map<String, MockTable> getTables()
     {
         return tables;
+    }
+
+    public Map<String, MockView> getViews()
+    {
+        return views;
     }
 
     public Map<String, String> getProperties()
@@ -52,6 +59,7 @@ public final class MockMapSchema
         return MoreObjects.toStringHelper(MockMapSchema.class.getSimpleName())
                 .add("name", name)
                 .add("tables", tables)
+                .add("views", views)
                 .add("properties", properties)
                 .toString();
     }
