@@ -10,6 +10,7 @@ import org.apache.arrow.vector.FixedSizeBinaryVector;
 import java.nio.charset.StandardCharsets;
 
 import static com.google.common.base.Verify.verify;
+import static java.lang.String.format;
 
 public interface ICharNWriter
 {
@@ -19,7 +20,7 @@ public interface ICharNWriter
     {
         String padded = rightPadSpaces(rawValue, typeLength);
         verify(CharMatcher.ascii().matchesAllOf(padded), "CHAR type supports only ASCII charset");
-        vector.set(getCount(), padded.getBytes(StandardCharsets.UTF_8));
+        vector.setSafe(getCount(), padded.getBytes(StandardCharsets.UTF_8));
     }
 
     default String rightPadSpaces(String value, int length)
@@ -28,13 +29,13 @@ public interface ICharNWriter
         // Spark does not support Char(N) so values are unpadded varchars, so pad short values and fail longer values
         // VAST requires padded values.
         if (length > value.length()) {
-            return String.format("%1$-" + length + "s", value);
+            return format("%1$-" + length + "s", value);
         }
         else if (length == value.length()) {
             return value;
         }
         else {
-            throw new IllegalArgumentException("Value too long");
+            throw new IllegalArgumentException(format("Value too long for type: CHAR(%s)", length));
         }
     }
 }
