@@ -8,8 +8,9 @@ import org.apache.spark.sql.connector.expressions.Expression;
 import org.apache.spark.sql.connector.expressions.Expressions;
 import org.apache.spark.sql.connector.expressions.NamedReference;
 import org.apache.spark.sql.connector.expressions.filter.Predicate;
-import org.apache.spark.sql.connector.read.colstats.ColumnStatistics;
 import org.apache.spark.sql.connector.read.Statistics;
+import org.apache.spark.sql.connector.read.colstats.ColumnStatistics;
+import org.apache.spark.sql.types.BinaryType;
 import org.apache.spark.sql.types.BooleanType;
 import org.apache.spark.sql.types.DataType;
 import org.apache.spark.sql.types.DateType;
@@ -20,12 +21,18 @@ import org.apache.spark.sql.types.StructType;
 import org.apache.spark.sql.types.TimestampType;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import scala.jdk.javaapi.OptionConverters;
 
 import java.util.Arrays;
+import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
+import java.util.Optional;
+import java.util.OptionalLong;
+import java.util.function.UnaryOperator;
 
-import scala.jdk.javaapi.OptionConverters;
+import static com.vastdata.OptionalPrimitiveHelpers.map;
+import static java.lang.Math.ceil;
 
 /* A simple statistics-based filter selectivity estimator, based on Spark's FilterEstimation.scala
  * TODO: Add hystogram-based estimation, once we have hystograms
@@ -33,57 +40,81 @@ import scala.jdk.javaapi.OptionConverters;
 public final class FilterEstimator
 {
     private static final Logger LOG = LoggerFactory.getLogger(FilterEstimator.class);
-    private static double estimateOpSelectivity(Predicate predicate, StructField field, NamedReference reference, Statistics statistics)
+
+    private FilterEstimator() {}
+
+    private static boolean isStringOrBinaryOrWithinRange(ColumnStatistics colStats, StructField field, Predicate predicate)
     {
-        ColumnStatistics colStats = statistics.columnStats().get(reference);
+        final ValueInterval statsInterval =
+            ValueInterval.apply(OptionConverters.toScala(colStats.min()), OptionConverters.toScala(colStats.max()), field.dataType());
+        final org.apache.spark.sql.connector.expressions.Literal jLiteral =
+            (org.apache.spark.sql.connector.expressions.Literal) predicate.children()[1];
+        final org.apache.spark.sql.catalyst.expressions.Literal sLiteral =
+                new org.apache.spark.sql.catalyst.expressions.Literal(jLiteral.value(), jLiteral.dataType());
+        return statsInterval.contains(sLiteral);
+    }
+
+    private static double estimateOpSelectivity(Predicate predicate, StructField field, NamedReference reference,
+                                                Statistics statistics, Map<NamedReference, ColumnStatistics> statsMap,
+                                                boolean updateStatistics)
+    {
+        final ColumnStatistics colStats =  statsMap.get(reference);
         if (colStats == null){
             return 1.0;
         }
+        final DataType dt = field.dataType();
         switch (predicate.name()) {
         case "<>":
         case "!=":
         case "=":
         {
-            if (!colStats.min().isPresent() || !colStats.max().isPresent()) {
-                return 1.0;
-            }
-            ValueInterval statsInterval =
-                ValueInterval.apply(OptionConverters.toScala(colStats.min()), OptionConverters.toScala(colStats.max()), field.dataType());
-            org.apache.spark.sql.connector.expressions.Literal jLiteral =
-                (org.apache.spark.sql.connector.expressions.Literal) predicate.children()[1];
-            org.apache.spark.sql.catalyst.expressions.Literal sLiteral = 
-                new org.apache.spark.sql.catalyst.expressions.Literal(jLiteral.value(), jLiteral.dataType());
             double percent = 1.0;
-            if (statsInterval.contains(sLiteral)) {
+            if (isStringOrBinaryOrWithinRange(colStats, field, predicate)) {
+                if (updateStatistics && "=".equals(predicate.name())) {
+                    ColumnStatistics newStat = null;
+                    if ((dt instanceof StringType) || (dt instanceof BinaryType)) {
+                        newStat =
+                            new ColumnLevelStatistics(OptionalLong.of(1), colStats.min(), colStats.max(),
+                                                      OptionalLong.of(0), colStats.avgLen(), colStats.maxLen());
+                    }
+                    else {
+                        final org.apache.spark.sql.connector.expressions.Literal jLiteral =
+                            (org.apache.spark.sql.connector.expressions.Literal) predicate.children()[1];
+                        newStat =
+                            new ColumnLevelStatistics(OptionalLong.of(1), Optional.of(jLiteral.value()),
+                                                      Optional.of(jLiteral.value()), OptionalLong.of(0),
+                                                      colStats.avgLen(), colStats.maxLen());
+                    }
+                    statsMap.replace(reference, newStat);
+                }
                 if (colStats.distinctCount().isPresent()) {
                     percent = 1.0/colStats.distinctCount().getAsLong();
                 }
-                percent = 1.0;
+                else {
+                    return 1.0;
+                }
             }
             else {
                 percent = 0.0;
             }
-            return predicate.name() == "="? percent : (1.0 - percent);
+            return "=".equals(predicate.name()) ? percent : (1.0 - percent);
         }
         case ">":
         case "<":
         case "<=":
         case ">=":
         {
-            DataType dt = field.dataType();
             // Non-numeric types???
             if (!((dt instanceof NumericType) || (dt instanceof DateType) || (dt instanceof TimestampType) || (dt instanceof BooleanType))  ||
                 !colStats.min().isPresent() || !colStats.max().isPresent() || !colStats.distinctCount().isPresent()) {
                 return 1.0;
             }
-            org.apache.spark.sql.connector.expressions.Literal literal =
+            final org.apache.spark.sql.connector.expressions.Literal literal =
                 (org.apache.spark.sql.connector.expressions.Literal) predicate.children()[1];
-            ValueInterval statsInterval =
-                ValueInterval.apply(OptionConverters.toScala(colStats.min()), OptionConverters.toScala(colStats.max()), dt);
-            double max = EstimationUtils.toDouble(colStats.max().get(), literal.dataType());
-            double min = EstimationUtils.toDouble(colStats.min().get(), literal.dataType());
-            long ndv = colStats.distinctCount().getAsLong();
-            double numericLiteral = EstimationUtils.toDouble(literal.value(), literal.dataType());
+            final double max = EstimationUtils.toDouble(colStats.max().get(), literal.dataType());
+            final double min = EstimationUtils.toDouble(colStats.min().get(), literal.dataType());
+            final long ndv = colStats.distinctCount().getAsLong();
+            final double numericLiteral = EstimationUtils.toDouble(literal.value(), literal.dataType());
             boolean noOverlap = false;
             boolean completeOverlap = false;
             switch (predicate.name()) {
@@ -148,6 +179,23 @@ public final class FilterEstimator
                 }
             }
             LOG.info("Estimating predicate: {}: name: {} min: {} max: {} selectivity: {}", predicate.toString(), predicate.name(), min, max, percent);
+            if (updateStatistics) {
+                Optional<Object> newMin = colStats.min();
+                Optional<Object> newMax = colStats.max();
+                switch (predicate.name()) {
+                case "<":
+                case "<=":
+                    newMax = Optional.of(literal.value());
+                    break;
+                case ">":
+                case ">=":
+                    newMin = Optional.of(literal.value());
+                }
+                ColumnStatistics newStat =
+                    new ColumnLevelStatistics(OptionalLong.of((long)ceil(ndv * percent)), newMin, newMax,
+                                              OptionalLong.of(0), colStats.avgLen(), colStats.maxLen());
+                statsMap.replace(reference, newStat);
+            }
             return percent;
         }
         case "IS_NULL":
@@ -156,16 +204,31 @@ public final class FilterEstimator
             if (!colStats.nullCount().isPresent()) {
                 return 1.0;
             }
-            long rowCount = statistics.numRows().getAsLong();
-            long nullCount = colStats.nullCount().getAsLong();
-            double nullPercent = rowCount == 0? 0.0 : (nullCount >= rowCount? 1.0 : (double) nullCount/(double) rowCount);
-            return predicate.name() == "IS_NULL"? nullPercent : (1.0 - nullPercent);
+            final long rowCount = statistics.numRows().getAsLong();
+            final long nullCount = colStats.nullCount().getAsLong();
+            final double nullPercent = rowCount == 0? 0.0 : (nullCount >= rowCount? 1.0 : (double) nullCount/(double) rowCount);
+            if (updateStatistics) {
+                ColumnStatistics newStat = null;
+                if ("IS_NULL".equals(predicate.name())) {
+                    newStat =
+                        new ColumnLevelStatistics(OptionalLong.of(0), Optional.empty(), Optional.empty(),
+                                                  colStats.nullCount(), colStats.avgLen(), colStats.maxLen());
+                }
+                else {
+                    newStat =
+                        new ColumnLevelStatistics(colStats.distinctCount(), colStats.min(), colStats.max(),
+                                                  OptionalLong.of(0), colStats.avgLen(), colStats.maxLen());
+                }
+                statsMap.replace(reference, newStat);
+            }
+            return "IS_NULL".equals(predicate.name()) ? nullPercent : (1.0 - nullPercent);
         }
         case "AND":
         {
-            Expression[] children = predicate.children();
+            final Expression[] children = predicate.children();
             if (children.length == 2 && (children[0] instanceof Predicate) && (children[1] instanceof Predicate)) {
-                return estimateOpSelectivity((Predicate) children[0], field, reference, statistics) * estimateOpSelectivity((Predicate) children[1], field, reference, statistics);
+                return estimateOpSelectivity((Predicate) children[0], field, reference, statistics, statsMap, updateStatistics)
+                    * estimateOpSelectivity((Predicate) children[1], field, reference, statistics, statsMap, updateStatistics);
             }
             return 1.0;
         }
@@ -174,10 +237,12 @@ public final class FilterEstimator
         }
     }
 
-    private static double estimateOrSelectivity(List<VastPredicate> predicates, Statistics statistics)
+    private static double estimateOrSelectivity(List<VastPredicate> predicates, Statistics statistics,
+                                                Map<NamedReference, ColumnStatistics> statsMap)
     {
         return predicates.stream()
-            .mapToDouble(p -> estimateOpSelectivity(p.getPredicate(), p.getField(), p.getReference(), statistics))
+            .mapToDouble(p -> estimateOpSelectivity(p.getPredicate(), p.getField(), p.getReference(), statistics,
+                                                    statsMap, predicates.size() == 1))
             .reduce((l, r) -> l + r - l * r)
             .orElse(1.0);
     }
@@ -186,16 +251,66 @@ public final class FilterEstimator
     {
         if (!statistics.numRows().isPresent())
             return 1.0;
+        if (!(statistics instanceof TableLevelStatistics)) {
+            LOG.warn("No statistics");
+            return 1.0;
+        }
+        Map<NamedReference, ColumnStatistics> statsMap =
+            new HashMap<NamedReference, ColumnStatistics>(((TableLevelStatistics)statistics).columnStats()); // shallow copy, because we might change it
         return predicates.stream()
-            .mapToDouble(l -> estimateOrSelectivity(l, statistics))
+            .mapToDouble(l -> estimateOrSelectivity(l, statistics, statsMap))
             .reduce(1.0, (l, r) -> l * r);
     }
 
+    private static Map<NamedReference, ColumnStatistics> updateColumnStatistics(final Map<NamedReference, ColumnStatistics> oldStats,
+                                                                                Map<NamedReference, ColumnStatistics> newStats,
+                                                                                double selectivity)
+    {
+        if (selectivity >= 1.0) {
+            return newStats;
+        }
+        for (Map.Entry<NamedReference, ColumnStatistics> pair : oldStats.entrySet()) {
+            ColumnStatistics updatedStats = newStats.get(pair.getKey());
+            OptionalLong dc = updatedStats.distinctCount();
+            if (dc.isPresent() && dc.getAsLong() > 1) {
+                dc = OptionalLong.of((long) ceil(dc.getAsLong() * selectivity));
+            }
+            OptionalLong nc = updatedStats.nullCount();
+            if (nc.isPresent() && nc.getAsLong() > 1) {
+                nc = OptionalLong.of((long) ceil(nc.getAsLong() * selectivity));
+            }
+            newStats.replace(pair.getKey(),
+                             new ColumnLevelStatistics(dc, updatedStats.min(), updatedStats.max(), nc,
+                                                       updatedStats.avgLen(), updatedStats.maxLen()));
+        }
+        return newStats;
+    }
+
+    public static TableLevelStatistics estimateStatistics(List<List<VastPredicate>> predicates,
+                                                          TableLevelStatistics statistics)
+    {
+        if (!statistics.numRows().isPresent())
+            return statistics;
+        if (!(statistics instanceof TableLevelStatistics)) {
+            LOG.warn("No statistics");
+            return statistics;
+        }
+        Map<NamedReference, ColumnStatistics> colStats =
+            new HashMap<NamedReference, ColumnStatistics>((statistics).columnStats()); // shallow copy, because we might change it
+        final double selectivity = predicates.stream()
+            .mapToDouble(l -> estimateOrSelectivity(l, statistics, colStats))
+            .reduce(1.0, (l, r) -> l * r);
+        final UnaryOperator<Long> applySelectivity = stat -> (long) ceil(stat * selectivity);
+        Map<NamedReference, ColumnStatistics> updatedColStats = updateColumnStatistics(statistics.columnStats(), colStats, selectivity);
+        return new TableLevelStatistics(map(statistics.sizeInBytes(), applySelectivity),
+                                        map(statistics.numRows(), applySelectivity),
+                                        updatedColStats);
+    }
 
     private static long fieldSize(StructField field, Map<NamedReference, ColumnStatistics> statsMap)
     {
-        NamedReference colRef = Expressions.column(field.name());
-        ColumnStatistics colStats = statsMap.get(colRef);
+        final NamedReference colRef = Expressions.column(field.name());
+        final ColumnStatistics colStats = statsMap.get(colRef);
         if (colStats == null || !colStats.avgLen().isPresent()) {
             LOG.debug("No statistics available for {}", field.name());
             return field.dataType().defaultSize();
@@ -204,6 +319,7 @@ public final class FilterEstimator
             return colStats.avgLen().getAsLong() + ((field.dataType() instanceof StringType)? (8 + 4) : 0);
         }
     }
+
     /*
      * Reimplementation of Sparks's EstimationUtils.getSizePerRow()
      */
@@ -212,6 +328,6 @@ public final class FilterEstimator
         return  8 +
             Arrays.stream(schema.fields())
             .mapToLong(f -> fieldSize(f, statistics.columnStats()))
-            .reduce(0, (l, r) -> l + r);
+            .reduce(0, Long::sum);
     }
-};
+}
