@@ -4,9 +4,11 @@
 
 package spark.sql.catalog.ndb;
 
+import com.google.common.collect.ImmutableList;
 import com.google.common.collect.ImmutableMap;
 import com.vastdata.client.VastClient;
 import com.vastdata.client.VastConfig;
+import com.vastdata.client.VastSchedulingInfo;
 import com.vastdata.client.error.VastConflictException;
 import com.vastdata.client.error.VastException;
 import com.vastdata.client.error.VastExceptionFactory;
@@ -15,28 +17,39 @@ import com.vastdata.client.error.VastUserException;
 import com.vastdata.client.schema.AlterTableContext;
 import com.vastdata.client.schema.CreateTableContext;
 import com.vastdata.client.schema.DropTableContext;
+import com.vastdata.client.schema.DropViewContext;
 import com.vastdata.client.schema.StartTransactionContext;
 import com.vastdata.client.schema.VastMetadataUtils;
 import com.vastdata.client.tx.SimpleVastTransaction;
+import com.vastdata.client.tx.VastTraceToken;
+import com.vastdata.client.tx.VastTransaction;
 import com.vastdata.client.tx.VastTransactionHandleManager;
+import com.vastdata.spark.SparkViewMetadata;
+import com.vastdata.spark.VastColumnarBatchReader;
+import com.vastdata.spark.VastInputPartition;
 import com.vastdata.spark.VastTable;
+import com.vastdata.spark.VastTableReadOnly;
+import com.vastdata.spark.VastView;
 import com.vastdata.spark.tx.VastAutocommitTransaction;
 import com.vastdata.spark.tx.VastSimpleTransactionFactory;
 import com.vastdata.spark.tx.VastSparkTransactionsManager;
 import ndb.DefaultSource;
 import ndb.NDB;
 import ndb.ka.NDBJobsListener;
-import net.bytebuddy.ClassFileVersion;
 import org.apache.arrow.vector.types.pojo.Field;
+import org.apache.spark.SparkConf;
 import org.apache.spark.SparkContext;
 import org.apache.spark.SparkContext$;
 import org.apache.spark.scheduler.SparkListenerInterface;
+import org.apache.spark.sql.catalyst.InternalRow;
 import org.apache.spark.sql.catalyst.analysis.NamespaceAlreadyExistsException;
 import org.apache.spark.sql.catalyst.analysis.NoSuchFunctionException;
 import org.apache.spark.sql.catalyst.analysis.NoSuchNamespaceException;
 import org.apache.spark.sql.catalyst.analysis.NoSuchTableException;
+import org.apache.spark.sql.catalyst.analysis.NoSuchViewException;
 import org.apache.spark.sql.catalyst.analysis.NonEmptyNamespaceException;
 import org.apache.spark.sql.catalyst.analysis.TableAlreadyExistsException;
+import org.apache.spark.sql.catalyst.analysis.ViewAlreadyExistsException;
 import org.apache.spark.sql.connector.catalog.CatalogExtension;
 import org.apache.spark.sql.connector.catalog.CatalogPlugin;
 import org.apache.spark.sql.connector.catalog.FunctionCatalog;
@@ -46,8 +59,10 @@ import org.apache.spark.sql.connector.catalog.Table;
 import org.apache.spark.sql.connector.catalog.TableChange;
 import org.apache.spark.sql.connector.catalog.functions.UnboundFunction;
 import org.apache.spark.sql.connector.expressions.Transform;
+import org.apache.spark.sql.types.StructField;
 import org.apache.spark.sql.types.StructType;
 import org.apache.spark.sql.util.CaseInsensitiveStringMap;
+import org.apache.spark.sql.vectorized.ColumnarBatch;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import scala.Option;
@@ -55,11 +70,13 @@ import spark.sql.catalog.ndb.alter.VastTableChange;
 import spark.sql.catalog.ndb.alter.VastTableChangeFactory;
 
 import java.util.Arrays;
+import java.util.Collections;
 import java.util.HashSet;
 import java.util.List;
 import java.util.Map;
 import java.util.Optional;
 import java.util.Set;
+import java.util.function.Supplier;
 import java.util.stream.Collectors;
 
 import static com.vastdata.client.ParsedURL.PATH_SEPERATOR;
@@ -69,28 +86,33 @@ import static com.vastdata.client.importdata.VastImportDataMetadataUtils.IMPORT_
 import static com.vastdata.client.importdata.VastImportDataMetadataUtils.IMPORT_DATA_TABLE_NAME_SUFFIX;
 import static com.vastdata.client.importdata.VastImportDataMetadataUtils.getTableNameForAPI;
 import static com.vastdata.client.importdata.VastImportDataMetadataUtils.isImportDataTableName;
+import static com.vastdata.client.schema.VastViewMetadata.COLUMN_ALIASES_FIELD;
+import static com.vastdata.client.schema.VastViewMetadata.COLUMN_COMMENTS_FIELD;
+import static com.vastdata.client.schema.VastViewMetadata.COMMENT_FIELD;
+import static com.vastdata.client.schema.VastViewMetadata.SQL_FIELD;
+import static com.vastdata.client.schema.VastViewMetadata.VIEW_METADATA_TABLE;
 import static java.lang.String.format;
+import static spark.sql.catalog.ndb.NDBRowLevelOperationIdentifier.isForRowLevelOp;
+import static spark.sql.catalog.ndb.NDBRowLevelOperationIdentifier.trimTableNameFromRowLevelOpSuffix;
 
 public class VastCatalog
         implements CatalogExtension
 {
     public static final String[] EMPTY_NAMESPACE = new String[0];
     private static final Logger LOG = LoggerFactory.getLogger(VastCatalog.class);
-    static {
-        ClassFileVersion classFileVersion = ClassFileVersion.ofThisVm();
-        LOG.info("Class file version of this vm: {}", classFileVersion);
-    }
+    public static final String[] DEFAULT_VAST_CATALOG = {"ndb"};
 
     private final static DefaultSource defaultVastSource = new DefaultSource();
 
     public static final int PAGE_SIZE = 1000; // TODO: use setting
+    private SparkConfValidator sparkConfValidator;
     private VastClient vastClient;
     private VastTransactionHandleManager<SimpleVastTransaction> transactionsManager;
 
     private FunctionCatalog functionsCatalogDelegate;
 
     @Override
-    public void initialize(String name, CaseInsensitiveStringMap options)
+    public void initialize(final String name, final CaseInsensitiveStringMap options)
     {
         LOG.debug("initialize {}, {}", name, options);
         try {
@@ -103,6 +125,8 @@ public class VastCatalog
             boolean empty = active.isEmpty();
             if (!empty) {
                 SparkContext sparkContext = active.get();
+                SparkConf conf = sparkContext.getConf();
+                sparkConfValidator = new SparkConfValidator(conf::getInt, conf::getBoolean);
                 Optional<SparkListenerInterface> any = sparkContext.listenerBus().listeners().stream().filter(l -> l instanceof NDBJobsListener).findAny();
                 if (!any.isPresent()) {
                     SparkListenerInterface instance = NDBJobsListener.instance(() -> vastClient, config);
@@ -117,6 +141,7 @@ public class VastCatalog
         catch (VastUserException e) {
             throw toRuntime(e);
         }
+        InitializedVastCatalog.setVastCatalog(this);
         LOG.debug("initialize {}, {}", name, options);
     }
 
@@ -130,7 +155,7 @@ public class VastCatalog
     public String[] defaultNamespace()
     {
         LOG.debug("defaultNamespace()");
-        return new String[] {"ndb"};
+        return DEFAULT_VAST_CATALOG;
     }
 
     @Override
@@ -305,7 +330,10 @@ public class VastCatalog
             LOG.debug("tableExists {} return {}", ident, exists);
             return exists;
         }
-        catch (Exception e) {
+        catch (final RuntimeException re) {
+            throw re;
+        }
+        catch (final Exception e) {
             throw new RuntimeException(format("Table existence check failed during fetching table info for identifier %s", ident.name()), e);
         }
     }
@@ -324,10 +352,14 @@ public class VastCatalog
             tableName = getTableNameForAPI(ident.name());
             LOG.debug("loadTable importing into table {}", tableName);
         }
+        else if (isForRowLevelOp(tableName)) {
+            tableName = trimTableNameFromRowLevelOpSuffix(ident.name());
+            LOG.debug("loadTable row level operation on table {}", tableName);
+        }
         try (VastAutocommitTransaction tx = VastAutocommitTransaction.wrap(vastClient, () -> transactionsManager.startTransaction(new StartTransactionContext(false, true)))) {
             Optional<String> vastTableHandleId = vastClient.getVastTableHandleId(tx, schemaName, tableName);
             if (vastTableHandleId.isPresent()) {
-                List<Field> fields = vastClient.listColumns(tx, schemaName, tableName, PAGE_SIZE);
+                List<Field> fields = vastClient.listColumns(tx, schemaName, tableName, PAGE_SIZE, Collections.emptyMap());
                 if (isImport) {
                     // Adjust schema of the table for only the fields the user mentioned as Spark is doing a strict validation
                     // Add at the end a field for the given imported filename
@@ -350,16 +382,16 @@ public class VastCatalog
                     fields.add(IMPORT_DATA_HIDDEN_FIELD);
                 }
                 StructType schema = TypeUtil.arrowFieldsListToSparkSchema(fields);
-                return new VastTable(schemaName, tableName, vastTableHandleId.get(), schema, () -> this.vastClient, isImport);
+                return makeVastTable(schemaName, tableName, vastTableHandleId.get(), schema, () -> this.vastClient, isImport);
             }
             else {
                 throw new NoSuchTableException(ident);
             }
         }
-        catch (NoSuchTableException noSuchTableException) {
-            throw noSuchTableException;
+        catch (final NoSuchTableException | RuntimeException rethrowable) {
+            throw rethrowable;
         }
-        catch (Exception e) {
+        catch (final Exception e) {
             throw new RuntimeException(format("Loading table failed during fetching table info for identifier %s", ident.name()), e);
         }
     }
@@ -379,13 +411,13 @@ public class VastCatalog
         List<Field> fieldList = TypeUtil.adaptVerifiedSparkSchemaToArrowFieldsList(schema);
         CreateTableContext ctx = new CreateTableContext(schemaName, tableName, fieldList, Optional.empty(), ImmutableMap.of());
 
-        try (VastAutocommitTransaction tx = VastAutocommitTransaction.wrap(vastClient, () -> transactionsManager.startTransaction(new StartTransactionContext(false, true)))) {
+        try (final VastAutocommitTransaction tx = VastAutocommitTransaction.wrap(vastClient, () -> transactionsManager.startTransaction(new StartTransactionContext(false, true)))) {
             if (vastClient.schemaExists(tx, schemaName)) {
                 if (!vastClient.tableExists(tx, schemaName, tableName)) {
                     vastClient.createTable(tx, ctx);
-                    String vastTableHandleId = vastClient.getVastTableHandleId(tx, schemaName, tableName).orElseThrow(() -> VastExceptionFactory.tableHandleIdNotFound(schemaName, tableName));
+                    final String vastTableHandleId = vastClient.getVastTableHandleId(tx, schemaName, tableName).orElseThrow(() -> VastExceptionFactory.tableHandleIdNotFound(schemaName, tableName));
                     tx.setCommit(true);
-                    return new VastTable(schemaName, tableName, vastTableHandleId, schema, () -> vastClient, false);
+                    return makeVastTable(schemaName, tableName, vastTableHandleId, schema, () -> vastClient, false);
                 }
                 else {
                     throw new TableAlreadyExistsException(ident);
@@ -395,10 +427,10 @@ public class VastCatalog
                 throw new NoSuchNamespaceException(ident.namespace());
             }
         }
-        catch (TableAlreadyExistsException | NoSuchNamespaceException sparkException) {
-            throw sparkException;
+        catch (final TableAlreadyExistsException | NoSuchNamespaceException | RuntimeException rethrowable) {
+            throw rethrowable;
         }
-        catch (Exception any) {
+        catch (final Exception any) {
             throw new RuntimeException(format("Creating table failed during putting table info to server for identifier %s", ident.name()), any);
         }
     }
@@ -422,17 +454,17 @@ public class VastCatalog
                 VastTableChange vastTableChange = vastTableChangeFactory.compose(changes);
                 vastTableChange.accept(vastClient, tx);
                 tx.setCommit(true);
-                List<Field> fields = vastClient.listColumns(tx, schemaName, tableName, PAGE_SIZE);
-                return new VastTable(schemaName, tableName, vastTableHandleId.get(), TypeUtil.arrowFieldsListToSparkSchema(fields), () -> vastClient, false);
+                List<Field> fields = vastClient.listColumns(tx, schemaName, tableName, PAGE_SIZE, Collections.emptyMap());
+                return makeVastTable(schemaName, tableName, vastTableHandleId.get(), TypeUtil.arrowFieldsListToSparkSchema(fields), () -> vastClient, false);
             }
             else {
                 throw new NoSuchTableException(ident);
             }
         }
-        catch (NoSuchTableException sparkException) {
-            throw sparkException;
+        catch (final NoSuchTableException | RuntimeException re) {
+            throw re;
         }
-        catch (Exception any) {
+        catch (final Exception any) {
             throw new RuntimeException(format("Failed applying table changes for identifier %s", ident.name()), any);
         }
     }
@@ -457,7 +489,10 @@ public class VastCatalog
                 return false;
             }
         }
-        catch (Exception e) {
+        catch (final RuntimeException re) {
+            throw re;
+        }
+        catch (final Exception e) {
             throw new RuntimeException(format("Failed dropping table for identifier %s", ident.name()), e);
         }
     }
@@ -473,7 +508,7 @@ public class VastCatalog
         String oldBucket = oldIdent.namespace()[0];
         String newBucket = newIdent.namespace()[0];
         if (!oldBucket.equalsIgnoreCase(newBucket)) {
-            throw new RuntimeException(format("Failed renaming table - changing bucket is not supported: %s, %s", oldBucket, newBucket));
+            throw new RuntimeException(format("Failed renaming table - changing bucket is not supported: %s, %s", oldIdent, newIdent));
         }
         String schemaName = compose(oldIdent.namespace());
         String tableName = oldIdent.name();
@@ -496,11 +531,160 @@ public class VastCatalog
             vastClient.alterTable(tx, schemaName, tableName, ctx);
             tx.setCommit(true);
         }
-        catch (NoSuchTableException | TableAlreadyExistsException sparkException) {
-            throw sparkException;
+        catch (final NoSuchTableException | TableAlreadyExistsException | RuntimeException rethrowable) {
+            throw rethrowable;
         }
-        catch (Exception e) {
+        catch (final Exception e) {
             throw new RuntimeException(format("Renaming table failed during update of table info for identifier %s", oldIdent), e);
+        }
+    }
+
+    private VastTable makeVastTable(String schemaName, String tableName, String handleID, StructType schema,
+                                    Supplier<VastClient> clientSupplier, boolean forImportData)
+    {
+        return sparkConfValidator.writeError
+                .map(error -> (VastTable) new VastTableReadOnly(schemaName, tableName, handleID, schema, clientSupplier, forImportData, error))
+                .orElseGet(() -> new VastTable(schemaName, tableName, handleID, schema, clientSupplier, forImportData));
+    }
+
+    public Identifier[] listViews(String... namespace) throws NoSuchNamespaceException {
+        LOG.debug("listViews {}", Arrays.toString(namespace));
+
+        if (namespace.length < 2) {
+            LOG.warn("Can't list views without specifying schema");
+            throw new NoSuchNamespaceException(namespace);
+        }
+
+        final String schemaName = compose(namespace);
+        try (final VastAutocommitTransaction tx = VastAutocommitTransaction.wrap(vastClient, () -> transactionsManager.startTransaction(new StartTransactionContext(false, true)))) {
+            if (!vastClient.schemaExists(tx, schemaName)) {
+                throw new NoSuchNamespaceException(namespace);
+            }
+            LOG.debug("Listing views for schema name: {}", schemaName);
+            try {
+                return vastClient.listViews(tx, schemaName, PAGE_SIZE).map(viewName -> Identifier.of(namespace, viewName)).toArray(Identifier[]::new);
+            }
+            catch (final VastServerException | VastUserException e) {
+                throw toRuntime(e);
+            }
+        }
+        catch (final VastException e) {
+            throw toRuntime(e);
+        }
+    }
+
+    public VastView loadView(final Identifier ident, Optional<VastTransaction> existingTransaction) throws NoSuchViewException
+    {
+        LOG.debug("loadViewSql {}", ident);
+        final String schemaName = compose(ident.namespace());
+        final String viewName = ident.name();
+        try (final VastAutocommitTransaction tx = VastAutocommitTransaction.wrap(existingTransaction, vastClient, () -> transactionsManager.startTransaction(new StartTransactionContext(false, true)))) {
+            if (vastClient.schemaExists(tx, schemaName)) {
+                if (vastClient.viewExists(tx, schemaName, viewName)) {
+                    VastTraceToken token = tx.generateTraceToken(Optional.of(format("getViewMetadata:%s", viewName)));
+                    VastConfig config = NDB.getConfig();
+                    VastSchedulingInfo schedulingInfo = vastClient.getSchedulingInfo(tx, token, schemaName, viewName);
+                    SimpleVastTransaction transaction = new SimpleVastTransaction(tx.getId(), tx.isReadOnly(), false);
+                    VastInputPartition partition = new VastInputPartition(null, 0, 0, 1);
+                    StructType structType = TypeUtil.arrowFieldsListToSparkSchema(ImmutableList.of(SQL_FIELD, COLUMN_ALIASES_FIELD, COLUMN_COMMENTS_FIELD, COMMENT_FIELD));
+                    Map<String, String> extraQueryParams = ImmutableMap.of("sub-table", VIEW_METADATA_TABLE);
+                    List<Field> fields = vastClient.listColumns(tx, schemaName, viewName, 1000, ImmutableMap.of());
+                    StructType viewSchema = new StructType(fields.stream().map(TypeUtil::arrowFieldToSparkField).toArray(StructField[]::new));
+                    try (VastColumnarBatchReader batchReader = new VastColumnarBatchReader(transaction, 0, config,
+                            schemaName, viewName, partition, structType, 1, Collections.emptyList(), schedulingInfo, false, extraQueryParams)) {
+                        while (batchReader.next()) {
+                            ColumnarBatch columnarBatch = batchReader.get();
+                            if (columnarBatch.numRows() > 0) {
+                                InternalRow row = columnarBatch.getRow(0);
+                                String sqlString = row.getUTF8String(0).toString();
+                                String[] aliases = rawObjectsArrayToStringsArray(row.getArray(1).array());
+                                String[] colComments = rawObjectsArrayToStringsArray(row.getArray(2).array());
+                                String comment = row.getString(3);
+                                return new VastView(viewName, sqlString, "ndb", comment, ident.namespace(), viewSchema, aliases, aliases, colComments);
+                            }
+                        }
+                        throw new RuntimeException("Failed to load view metadata " + ident);
+                    }
+
+                }
+                else {
+                    throw new NoSuchViewException(ident);
+                }
+            }
+            else {
+                throw new NoSuchViewException(ident);
+            }
+        }
+        catch (final NoSuchViewException | RuntimeException rethrowable) {
+            throw rethrowable;
+        }
+        catch (final Exception any) {
+            throw new RuntimeException(format("Loading view failed for identifier %s", viewName), any);
+        }
+    }
+
+    private static String[] rawObjectsArrayToStringsArray(Object[] rawAliasArray)
+    {
+        return rawAliasArray == null ? new String[0] : Arrays.stream(rawAliasArray).map(o -> o == null ? null : o.toString()).toArray(String[]::new);
+    }
+
+    public void createView(SparkViewMetadata ctx, boolean replace, Optional<VastTransaction> existingTransaction)
+            throws ViewAlreadyExistsException, NoSuchNamespaceException
+    {
+        LOG.debug("createView: CreateSparkViewContext: {}", ctx);
+        final String schemaName = compose(ctx.getIdentifier().namespace());
+        final String viewName = ctx.getIdentifier().name();
+        try (final VastAutocommitTransaction tx = VastAutocommitTransaction.wrap(existingTransaction, vastClient, () -> transactionsManager.startTransaction(new StartTransactionContext(false, true)))) {
+            if (vastClient.schemaExists(tx, schemaName)) {
+                if (!vastClient.viewExists(tx, schemaName, viewName)) {
+                    vastClient.createView(tx, ctx.toVastCreateViewContext());
+                }
+                else {
+                    if (replace) {
+                        LOG.debug("createView: replacing existing view");
+                        vastClient.dropView(tx, new DropViewContext(schemaName, viewName));
+                        vastClient.createView(tx, ctx.toVastCreateViewContext());
+                    }
+                    else {
+                        throw new ViewAlreadyExistsException(ctx.getIdentifier());
+                    }
+                }
+            }
+            else {
+                throw new NoSuchNamespaceException(ctx.getIdentifier().namespace());
+            }
+        }
+        catch (final ViewAlreadyExistsException | NoSuchNamespaceException | RuntimeException re) {
+            throw re;
+        }
+        catch (final Exception any) {
+            throw new RuntimeException(format("Creating view failed during putting view info to server for identifier %s", viewName), any);
+        }
+    }
+
+    public boolean dropView(Identifier ident, Optional<VastTransaction> existingTransaction) {
+        LOG.debug("dropView {}", ident);
+
+        final String schemaName = compose(ident.namespace());
+        final String viewName = ident.name();
+
+        final DropViewContext ctx = new DropViewContext(schemaName, viewName);
+
+        try (final VastAutocommitTransaction tx = VastAutocommitTransaction.wrap(existingTransaction, vastClient, () -> transactionsManager.startTransaction(new StartTransactionContext(false, true)))) {
+            if (vastClient.viewExists(tx, schemaName, viewName)) {
+                vastClient.dropView(tx, ctx);
+                tx.setCommit(true);
+                return true;
+            }
+            else {
+                return false;
+            }
+        }
+        catch (final RuntimeException re) {
+            throw re;
+        }
+        catch (final Exception e) {
+            throw new RuntimeException(format("Failed dropping view for identifier %s", ident.name()), e);
         }
     }
 }
