@@ -80,9 +80,9 @@ public class TestQueryDataResponseSchemaConstructor
         assertEquals(unit.childToParent, expectedReverseMapping);
         Block booleanBlock = wrapByteArrayAsBooleanBlockWithoutNulls(new byte[1]);
         ShortArrayBlock smallIntBlock = new ShortArrayBlock(1, Optional.of(new boolean[] {false}), new short[] {4});
-        Block smallIntRowBlock = RowBlock.fromFieldBlocks(1, Optional.of(new boolean[] {false}), new Block[] {RowBlock.fromFieldBlocks(1, Optional.of(new boolean[] {false}), new Block[] {smallIntBlock})});
+        Block smallIntRowBlock = RowBlock.fromNotNullSuppressedFieldBlocks(1, Optional.of(new boolean[] {false}), new Block[] {RowBlock.fromNotNullSuppressedFieldBlocks(1, Optional.of(new boolean[] {false}), new Block[] {smallIntBlock})});
         ByteArrayBlock byteIntBlock = new ByteArrayBlock(1, Optional.of(new boolean[] {false}), new byte[] {3});
-        Block tinyIntRowBlock = RowBlock.fromFieldBlocks(1, Optional.of(new boolean[] {false}), new Block[] {RowBlock.fromFieldBlocks(1, Optional.of(new boolean[] {false}), new Block[] {byteIntBlock})});
+        Block tinyIntRowBlock = RowBlock.fromNotNullSuppressedFieldBlocks(1, Optional.of(new boolean[] {false}), new Block[] {RowBlock.fromNotNullSuppressedFieldBlocks(1, Optional.of(new boolean[] {false}), new Block[] {byteIntBlock})});
         Slice slice = Slices.utf8Slice("somestring");
         int offset = slice.length();
         VariableWidthBlock varcharBlock = new VariableWidthBlock(1, slice, new int[] {0, offset}, Optional.of(new boolean[] {false}));
@@ -125,9 +125,9 @@ public class TestQueryDataResponseSchemaConstructor
         Slice slice = Slices.utf8Slice("somestring");
         int offset = slice.length();
         VariableWidthBlock varcharBlock = new VariableWidthBlock(1, slice, new int[] {0, offset}, Optional.of(new boolean[] {false}));
-        Block row1 = RowBlock.fromFieldBlocks(1, Optional.of(new boolean[]{false}), new Block[] {intBlock});
-        Block row2 = RowBlock.fromFieldBlocks(1, Optional.of(new boolean[]{false}), new Block[] {booleanBlock});
-        Block row3 = RowBlock.fromFieldBlocks(1, Optional.of(new boolean[]{false}), new Block[] {varcharBlock});
+        Block row1 = RowBlock.fromNotNullSuppressedFieldBlocks(1, Optional.of(new boolean[] {false}), new Block[] {intBlock});
+        Block row2 = RowBlock.fromNotNullSuppressedFieldBlocks(1, Optional.of(new boolean[] {false}), new Block[] {booleanBlock});
+        Block row3 = RowBlock.fromNotNullSuppressedFieldBlocks(1, Optional.of(new boolean[] {false}), new Block[] {varcharBlock});
         Block[] mockBlocks = new Block[] {row1, row2, row3};
         Page constructedPage = unit.construct(mockBlocks, 0);
         assertEquals(constructedPage.getChannelCount(), 1);
@@ -181,7 +181,7 @@ public class TestQueryDataResponseSchemaConstructor
         LongArrayBlock longBlock = new LongArrayBlock(1, Optional.of(new boolean[] {false}), new long[] {1});
         Block arrayBlock = ArrayBlock.fromElementBlock(1, Optional.of(new boolean[] {false}), new int[] {0, 1}, longBlock);
         Block[] mockBlocks = new Block[] {intBlock, booleanBlock, varcharBlock, arrayBlock,
-                RowBlock.fromFieldBlocks(1, Optional.of(new boolean[]{false}), new Block[] {arrayBlock})};
+                RowBlock.fromNotNullSuppressedFieldBlocks(1, Optional.of(new boolean[] {false}), new Block[] {arrayBlock})};
         Page constructedPage = unit.construct(mockBlocks, 0);
         assertEquals(constructedPage.getChannelCount(), 5);
         assertTrue(constructedPage.getBlock(3) instanceof ArrayBlock);
@@ -227,14 +227,14 @@ public class TestQueryDataResponseSchemaConstructor
         assertEquals(unit.childToParent, expectedReverseMapping);
         LongArrayBlock longBlock = new LongArrayBlock(1, Optional.of(new boolean[] {false}), new long[] {1});
         Block bigintArrayBlock = ArrayBlock.fromElementBlock(1, Optional.of(new boolean[] {false}), new int[] {0, 1}, longBlock);
-        Block projection2 = RowBlock.fromFieldBlocks(1, Optional.of(new boolean[] {false}), new Block[] {bigintArrayBlock});
+        Block projection2 = RowBlock.fromNotNullSuppressedFieldBlocks(1, Optional.of(new boolean[] {false}), new Block[] {bigintArrayBlock});
         Block boolBlock = new ByteArrayBlock(1, Optional.of(new boolean[] {false}), new byte[] {1});
-        Block projection3 = RowBlock.fromFieldBlocks(1, Optional.of(new boolean[] {false}), new Block[] {boolBlock});
+        Block projection3 = RowBlock.fromNotNullSuppressedFieldBlocks(1, Optional.of(new boolean[] {false}), new Block[] {boolBlock});
         Block tinyintArrayBlock = ArrayBlock.fromElementBlock(1, Optional.of(new boolean[] {false}), new int[] {0, 1}, boolBlock);
-        Block subRowArrayBlock = RowBlock.fromFieldBlocks(1, Optional.of(new boolean[]{false}), new Block[] {tinyintArrayBlock});
-        Block subRowBoolBlock = RowBlock.fromFieldBlocks(1, Optional.of(new boolean[]{false}), new Block[] {boolBlock});
-        Block projection5 = RowBlock.fromFieldBlocks(1, Optional.of(new boolean[] {false}), new Block[] {subRowBoolBlock});
-        Block projection7 = RowBlock.fromFieldBlocks(1, Optional.of(new boolean[] {false}), new Block[] {subRowArrayBlock});
+        Block subRowArrayBlock = RowBlock.fromNotNullSuppressedFieldBlocks(1, Optional.of(new boolean[] {false}), new Block[] {tinyintArrayBlock});
+        Block subRowBoolBlock = RowBlock.fromNotNullSuppressedFieldBlocks(1, Optional.of(new boolean[] {false}), new Block[] {boolBlock});
+        Block projection5 = RowBlock.fromNotNullSuppressedFieldBlocks(1, Optional.of(new boolean[] {false}), new Block[] {subRowBoolBlock});
+        Block projection7 = RowBlock.fromNotNullSuppressedFieldBlocks(1, Optional.of(new boolean[] {false}), new Block[] {subRowArrayBlock});
         Block[] testBlocks = new Block[] {projection2, projection3, projection5, projection7};
         Page constructedPage = unit.construct(testBlocks, 1);
         assertEquals(constructedPage.getChannelCount(), 1);
@@ -292,22 +292,22 @@ public class TestQueryDataResponseSchemaConstructor
         ByteArrayBlock byteBlock = new ByteArrayBlock(1, Optional.of(new boolean[] {false}), new byte[] {1});
         Block byteArrayBlock = ArrayBlock.fromElementBlock(1, Optional.of(new boolean[] {false}), new int[] {0, 1}, byteBlock);
         LongArrayBlock longBlock = new LongArrayBlock(1, Optional.of(new boolean[] {false}), new long[] {1});
-        Block nestedRowSubBlock1 = RowBlock.fromFieldBlocks(1, Optional.of(new boolean[] {false}), new Block[] {byteBlock});
-        Block nestedRow1 = RowBlock.fromFieldBlocks(1, Optional.of(new boolean[] {false}), new Block[] {nestedRowSubBlock1});
-        Block nestedRowSubBlock2 = RowBlock.fromFieldBlocks(1, Optional.of(new boolean[] {false}), new Block[] {shortArrayBlock});
-        Block nestedRow2 = RowBlock.fromFieldBlocks(1, Optional.of(new boolean[] {false}), new Block[] {nestedRowSubBlock2});
-        Block nestedrow3 = RowBlock.fromFieldBlocks(1, Optional.of(new boolean[] {false}), new Block[] {longBlock});
-        Block nestedrow4 = RowBlock.fromFieldBlocks(1, Optional.of(new boolean[] {false}), new Block[] {byteArrayBlock});
+        Block nestedRowSubBlock1 = RowBlock.fromNotNullSuppressedFieldBlocks(1, Optional.of(new boolean[] {false}), new Block[] {byteBlock});
+        Block nestedRow1 = RowBlock.fromNotNullSuppressedFieldBlocks(1, Optional.of(new boolean[] {false}), new Block[] {nestedRowSubBlock1});
+        Block nestedRowSubBlock2 = RowBlock.fromNotNullSuppressedFieldBlocks(1, Optional.of(new boolean[] {false}), new Block[] {shortArrayBlock});
+        Block nestedRow2 = RowBlock.fromNotNullSuppressedFieldBlocks(1, Optional.of(new boolean[] {false}), new Block[] {nestedRowSubBlock2});
+        Block nestedrow3 = RowBlock.fromNotNullSuppressedFieldBlocks(1, Optional.of(new boolean[] {false}), new Block[] {longBlock});
+        Block nestedrow4 = RowBlock.fromNotNullSuppressedFieldBlocks(1, Optional.of(new boolean[] {false}), new Block[] {byteArrayBlock});
         Page constructedPage = unit.construct(new Block[] {arraycolBlock, nestedRow1, nestedRow2, nestedrow3, nestedrow4}, 1);
         assertEquals(constructedPage.getChannelCount(), 2);
         assertTrue(constructedPage.getBlock(0) instanceof ArrayBlock);
         Block rowBlock = constructedPage.getBlock(1);
         assertTrue(rowBlock instanceof RowBlock);
-        List<Block> children = rowBlock.getChildren();
+        List<Block> children = ((RowBlock) rowBlock).getFieldBlocks();
         assertEquals(children.size(), 3);
         Block nestedRow = children.get(0);
         assertTrue(nestedRow instanceof RowBlock);
-        List<Block> grandChildren = nestedRow.getChildren();
+        List<Block> grandChildren = ((RowBlock) nestedRow).getFieldBlocks();
         assertEquals(grandChildren.size(), 2);
         Block boolBlock = grandChildren.get(0);
         Block arrayBlock = grandChildren.get(1);
@@ -352,22 +352,23 @@ public class TestQueryDataResponseSchemaConstructor
         assertEquals(unit.childToParent, expectedReverseMapping);
 
         ByteArrayBlock byteBlock = new ByteArrayBlock(1, Optional.of(new boolean[] {false}), new byte[] {1});
-        Block nestedRowSubBlock1 = RowBlock.fromFieldBlocks(1, Optional.of(new boolean[] {false}), new Block[] {byteBlock});
-        Block nestedRow1 = RowBlock.fromFieldBlocks(1, Optional.of(new boolean[] {false}), new Block[] {nestedRowSubBlock1});
+        Block nestedRowSubBlock1 = RowBlock.fromNotNullSuppressedFieldBlocks(1, Optional.of(new boolean[] {false}), new Block[] {byteBlock});
+        Block nestedRow1 = RowBlock.fromNotNullSuppressedFieldBlocks(1, Optional.of(new boolean[] {false}), new Block[] {nestedRowSubBlock1});
 
         LongArrayBlock longBlock = new LongArrayBlock(1, Optional.of(new boolean[] {false}), new long[] {1});
         Block longArrayBlock = ArrayBlock.fromElementBlock(1, Optional.of(new boolean[] {false}), new int[] {0, 1}, longBlock);
-        Block nestedRowSubBlock2 = RowBlock.fromFieldBlocks(1, Optional.of(new boolean[] {false}), new Block[] {longArrayBlock});
-        Block nestedRow2 = RowBlock.fromFieldBlocks(1, Optional.of(new boolean[] {false}), new Block[] {nestedRowSubBlock2});
+        Block nestedRowSubBlock2 = RowBlock.fromNotNullSuppressedFieldBlocks(1, Optional.of(new boolean[] {false}), new Block[] {longArrayBlock});
+        Block nestedRow2 = RowBlock.fromNotNullSuppressedFieldBlocks(1, Optional.of(new boolean[] {false}), new Block[] {nestedRowSubBlock2});
 
         Page constructedPage = unit.construct(new Block[] {nestedRow1, nestedRow2}, 1);
         assertEquals(constructedPage.getChannelCount(), 1);
         Block rowBlock = constructedPage.getBlock(0);
         assertTrue(rowBlock instanceof RowBlock);
-        List<Block> children = rowBlock.getChildren();
+        List<Block> children = ((RowBlock) rowBlock).getFieldBlocks();
         assertEquals(children.size(), 1);
         Block childRowBlock = children.get(0);
-        List<Block> grandChildren = childRowBlock.getChildren();
+        assertTrue(childRowBlock instanceof RowBlock);
+        List<Block> grandChildren = ((RowBlock) childRowBlock).getFieldBlocks();
         assertEquals(grandChildren.size(), 2);
         Block boolBlock = grandChildren.get(0);
         Block arrayBlock = grandChildren.get(1);
